// src/app/Feature/Motobike/models/motorbike.model.ts
export interface Motorbike{
  id: string
  name: string
  type: number
  speed: number
  capacity: number
  color: string
  yearOfManufacture: number
  madeIn: string
  status: number
  description: string
  priceDay: number
  priceWeek: number
  priceMonth: number
  motorbikeAvatar: string
  licensePlate: string
}